+++ src/Urls/ArchivedUrlException.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace Thinktomorrow\Chief\Urls;
+
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+
+class ArchivedUrlException extends NotFoundHttpException
+{
+    //
+}